C - Function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,5 @@
 #include "function_pointers.h"
 #include <stdio.h>
-#include <stdlib.h>
 
 /**
  * array_iterator - executes a function given as a parameter
@@ -13,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	unsigned int i;
 
 	if (array && action)
 	{
